feat: add snapshot test
